moved logout button under navbar
